% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{chipSubset}
\alias{chipSubset}
\title{ChIP-seq bam file stored as spp tag-list for 
a subset of chromosomes for the chip}
\format{list of 2 elements containing the reads and the 
read quality of the ChIP. 

\itemize{
    \item tags : list containing the start coordinates 
of each read aligned (ChIP) (3'end) 
    \item quality: list containing the read quality 
of each read
}}
\usage{
data(chipSubset)
}
\description{
Example data for manual and vignette. 
tag-list created with the
read.bam.tags() function from spp package.
The original bam file has been dowloaded from ENCODE 
(ID: ENCFF000BFX).
}
\keyword{bamfile}
