% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{chr2Bam}
\alias{chr2Bam}
\title{ChIP-seq bam file stored as spp tag-list chr2}
\format{list of 2 elements containing the reads and the 
read quality of the ChIP. 

\itemize{
    \item tags : list containing the start coordinates 
of each read aligned to chrom2 (ChIP) (3'end) 
    \item quality: list containing the read quality 
of each read
}}
\usage{
data(chr2Bam)
}
\description{
Example data for manual and vignette. 
chr1 tag-list created with the
read.bam.tags() function from spp package.
The original bam file has been dowloaded from ENCODE 
(ID: ENCFF000BFX).
}
\keyword{bamfile}
