% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{classesDefList}
\alias{classesDefList}
\title{List with definitions}
\format{
A list of 5, containing a vector with chromatin mark names, 
TF names, the members of sharp binding marks, broad binding marks 
and RNApol2

\itemize{
    \item Hlist: vector with chromatin marks
    \item allSharp: vector with sharp binding marks
    \item allBroad: vector  with broad binding marks
    \item RNAPol2 : RNApol2
    \item TF: vector with TF names
}
}
\usage{
data(classesDefList)
}
\description{
List containing information about chromatin marks, transcription
factors, definitions for broad and sharp class.
}
\keyword{definitions}
