% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_db}
\alias{compendium_db}
\title{Histone mark compendium}
\format{Data frame with 366 variables (quality control metrics, metadata) for each 
analysed sample.}
\source{
XX
}
\usage{
data(compendium_db)
}
\description{
Histone mark ChIP-seq compendium for the CHIC package.
The compendium contains quality control metrics and metadata for 2329 histone mark samples 
analysed from ENCODE and Roadmap Epigenomics.
}
\keyword{compendium}
\keyword{control}
\keyword{histone}
\keyword{marks}
\keyword{metrics}
\keyword{quality}
