% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_db_tf}
\alias{compendium_db_tf}
\title{Transcription factor compendium}
\format{A data frame with 366 variables to describe each ENCODE experiment.}
\source{

}
\usage{
data(compendium_db_tf)
}
\description{
Transcription factor ChIP-seq compendium for the CHIC package.
}
\keyword{tf}
