% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_profiles}
\alias{compendium_profiles}
\title{Metagene profile data.}
\format{A list of data frames with variables:}
\source{

}
\usage{
compendium_profiles
}
\description{
Metagene profile data for the CHIC package.
}
\examples{


}
\keyword{datasets}
