% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_profiles_TF}
\alias{compendium_profiles_TF}
\title{Metagene profile data for transcription factors}
\format{A list of data frames with the coordinates of the 
metagene profile of the respective transcription factor.

    x : genomic coordinates
    mean : the mean of the signal intensity in the compendium
    sd : the standard deviation of the signal intensity in the 
    compendium
    q1..q5 : being the respective quantile of the value distribution
    sderr : standard error}
\source{
XX
}
\usage{
data(compendium_profiles_TF)
}
\description{
Compendium of averaged metagene profiles for the ChIC package. Contains 
averaged metagene profiles for following ChIP-seq datasets from ENCODE.
}
\details{
\itemize{
    \item ARID3A
    \item ATF2
    \item ATF3
    \item BACH1
    \item BATF
    \item BCL11A
    \item BCL3
    \item BCLAF1
    \item BHLHE40
    \item BRCA1
    \item CBX2
    \item CBX3
    \item CEBPB
    \item CEBPD
    \item CEBPZ
    \item CHD1
    \item CHD2
    \item CHD4
    \item CREB1
    \item CREBBP
    \item CTCF
    \item CUX1
    \item E2F4
    \item E2F6
    \item EBF1
    \item EGR1
    \item ELF1
    \item ELK1
    \item EP300
    \item ESRRA
    \item ETS1
    \item EZH2
    \item FOSL1
    \item FOSL2
    \item FOS
    \item FOXA1
    \item FOXA2
    \item FOXM1
    \item GABPA
    \item GATA2
    \item GATA3
    \item GTF2F1
    \item HA
    \item HDAC2
    \item HDAC6
    \item HNF4A
    \item HNF4G
    \item IRF3
    \item IRF4
    \item JUND
    \item JUN
    \item KAT2A
    \item KAT2B
    \item KDM1A
    \item KDM4A
    \item KDM5A
    \item KDM5B
    \item MAFF
    \item MAFK
    \item MAX
    \item MAZ
    \item MBD4
    \item MEF2A
    \item MEF2C
    \item MTA3
    \item MXI1
    \item MYBL2
    \item MYC
    \item NANOG
    \item NCOR1
    \item NFATC1
    \item NFIC
    \item NR2F2
    \item NR3C1
    \item NRF1
    \item PAX5
    \item PBX3
    \item PHF8
    \item PML
    \item POU2F2
    \item PRDM1
    \item RAD21
    \item RBBP5
    \item RCOR1
    \item REST
    \item RFX5
    \item RNF2
    \item RUNX3
    \item RXRA
    \item SAP30
    \item SETDB1
    \item SIN3A
    \item SIRT6
    \item SIX5
    \item SMC3
    \item SP1
    \item SP2
    \item SP4
    \item SPI1
    \item SREBF1
    \item SREBF2
    \item SRF
    \item STAT3
    \item STAT5A
    \item SUPT20H
    \item SUZ12
    \item TAF1
    \item TAF7
    \item TAL1
    \item TBP
    \item TCF12
    \item TCF3
    \item TEAD4
    \item USF1
    \item USF2
    \item WRNIP1
    \item YY1
    \item ZBTB33
    \item ZBTB7A
    \item ZC3H11A
    \item ZEB1
    \item ZMIZ1
    \item ZNF143
    \item ZNF274
    \item ZNF384
}
}
\keyword{Transcription}
\keyword{compendium}
\keyword{factors}
\keyword{metagene}
\keyword{profiles}
