% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{crossvalues_Chip}
\alias{crossvalues_Chip}
\title{CrossCorrelation values for example ChIP-seq data for Vignette}
\format{list of 20 elements containing EM scores}
\usage{
data(crossvalues_Chip)
}
\description{
Example data to be used in the vignette for chrom2
}
\keyword{crossvalues_Chip}
