% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{dm3_chrom_info}
\alias{dm3_chrom_info}
\title{dm3 chromosome information}
\format{
A named list of int vectors with 2 elements, the start and end
position of all dm3 chromosomes.
}
\source{
http://hgdownload.cse.ucsc.edu/goldenPath/dm3/database/
chromInfo.txt.gz
}
\usage{
data(dm3_chrom_info)
}
\description{
dm3 chromosome information for the ChIC package.
}
\keyword{dm3}
