% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{dm6_chrom_info}
\alias{dm6_chrom_info}
\title{dm6 chromosome information}
\format{A named list of int vectors with 2 elements, the start and end
position of all dm6 chromosomes.}
\source{
http://hgdownload.cse.ucsc.edu/goldenPath/dm6/database/
chromInfo.txt.gz
}
\usage{
data(dm6_chrom_info)
}
\description{
dm6 chromosome information for the ChIC package.
}
\keyword{dm6}
