% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{EM_Results}
\alias{EM_Results}
\title{Example results from the qualityScores_EM() function based on the example ChIP-seq data for the Vignette for ChIC package}
\format{
list containing EM scores and filtered reads
}
\usage{
data(EM_Results)
}
\description{
Example data to be used in the vignette for ChIC package
}
\keyword{EM_Results}
