% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{EM_scores}
\alias{EM_scores}
\title{EM values for example ChIP-seq data for Vignette}
\format{list of 5 elements containing EM scores

\itemize{
    \item QCscores_ChIP : QCscores for ChIP
    \item QCscores_Input: QCscores for Input
    \item QCscores_binding: QCscores from peak calling
}}
\usage{
data(EM_scores)
}
\description{
Example data to be used in the vignette.
}
\keyword{EM_scores}
