% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{hg38_chrom_info}
\alias{hg38_chrom_info}
\title{hg38 chromosome information}
\format{
A named list of int vectors with 2 elements, the start and end
position of all hg38 chromosomes.
}
\source{
http://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/
hg38.chrom.sizes
}
\usage{
data(hg38_chrom_info)
}
\description{
hg38 chromosome information for the ChIC package.
}
\keyword{hg38}
