% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{inputBam}
\alias{inputBam}
\title{ChIP-seq bam file stored as spp tag-list}
\format{list of 2 elements containing the reads and the 
read quality. 

\itemize{
    \item tags : list of chromosomes containing 
the start coordinate of each read (3'end) 
    \item quality: list of chromosomes containing the 
read quality of each read
}}
\usage{
data(chipBam)
}
\description{
Example data for manual and vignette. 
Tag-list file created with the
read.bam.tags() function from spp package.
The original bam file has been dowloaded from ENCODE 
(ID: ENCFF000BLL).
}
\keyword{bamfile}
