% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{inputSubset}
\alias{inputSubset}
\title{ChIP-seq bam file stored as spp tag-list for 
a subset of chromosomes for the input}
\format{
list of 2 elements containing the reads and the 
read quality of the input. 

\itemize{
    \item tags : list containing the start coordinates 
of each read aligned (input) (3'end) 
    \item quality: list containing the read quality 
of each read
}
}
\usage{
data(inputSubset)
}
\description{
Example data for manual and vignette. 
Tag-list created with the
read.bam.tags() function from spp package for 
the input data.
The original bam file has been dowloaded from ENCODE 
(ID: ENCFF000BDQ).
}
\keyword{bamfile}
