% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{mm10_chrom_info}
\alias{mm10_chrom_info}
\title{mm10 chromosome information}
\format{
A named list of int vectors with 2 elements, the start and end
position of all mm10 chromosomes.
}
\source{
http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips/
mm10.chrom.sizes
}
\usage{
data(mm10_chrom_info)
}
\description{
mm10 chromosome information for the ChIC package.
}
\keyword{mm10}
