% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{mm9_refseq_genes_filtered_granges}
\alias{mm9_refseq_genes_filtered_granges}
\title{Filtered RefSeq gene annotation as GRanges for mm9}
\format{
A GRanges object.
}
\usage{
data(mm9_refseq_genes_filtered_granges)
}
\description{
RefSeq gene annotation data for the ChIC package in GRanges format, 
filtered by the gene length and by overlaps.
}
\keyword{annotation}
