% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{refseq_genes_filtered_granges}
\alias{refseq_genes_filtered_granges}
\title{Filtered RefSeq gene annotation as GRanges}
\format{A GRanges object.}
\usage{
data(refseq_genes_filtered_granges)
}
\description{
RefSeq gene annotation data for the CHIC package in GRanges format, filtered
by ...
}
\keyword{annotation}
