% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{rf_models}
\alias{rf_models}
\title{Random forest models for chromatin marks ChIP-seq 
experiment classification}
\format{A list of 7 random forest models for the different 
chromatin marks and transcription factors:

\itemize{
    \item broadEncode : model for broad binding marks
    \item H3K9Encode : model for H3K9me3
    \item H3K27Encode : model for H3K27me3
    \item H3K36Encode : model for H3K36me3
    \item RNAPol2Encode : model for RNAPol2
    \item sharpEncode : model for sharp binding marks
    \item TFmodel : model for transcription factors
}}
\usage{
data(rf_models)
}
\description{
Random forest models based on ENCODE and Roadmap data for ChIP-seq 
experiment classification using the ChIC package.
}
\keyword{classification}
