% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{tf_model}
\alias{tf_model}
\title{Random forest model for transcription factor ChIP-seq experiment
classification}
\format{A single random forest model.}
\usage{
data(tf_model)
}
\description{
Random forest model based on ENCODE data for transcription factor ChIP-seq
experiment classification using the CHIC package.
}
\keyword{classification}
