% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_db}
\alias{compendium_db}
\title{Histone mark compendium}
\format{A data frame with 366 variables to describe each ENCODE or Roadmap
experiment.}
\source{

}
\usage{
data(compendium_db)
}
\description{
Histone mark ChIP-seq compendium for the CHIC package.
}
\keyword{histone}
