% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{compendium_profiles}
\alias{compendium_profiles}
\title{Metagene profile data}
\format{A list of data frames with the coordinates of the 
metagene profile of the respective chromatin mark.

    x : genomic coordinates
    mean : the mean of the signal intensity in the compendium
    sd : the standard deviation of the signal intensity in the 
    compendium
    q1..q5 : being the respective quantile of the value distribution
    sderr : standard error}
\source{
XX
}
\usage{
data(compendium_profiles)
}
\description{
Compendium of averaged metagene profiles for the ChIC package. Contains 
averaged metagene profiles for following ChIP-seq datasets from ENCODE 
and Roadmap Epigenomics.
}
\details{
\itemize{
    \item H2A.Z
    \item H2AFZ
    \item H2AK5ac
    \item H2AK9ac
    \item H2BK120ac
    \item H2BK12ac
    \item H2BK15ac
    \item H2BK20ac
    \item H2BK5ac
    \item H3K14ac
    \item H3K18ac
    \item H3K23ac
    \item H3K23me2
    \item H3K27ac
    \item H3K27me3
    \item H3K36me3
    \item H3K4ac
    \item H3K4me1
    \item H3K4me2
    \item H3K4me3
    \item H3K56ac
    \item H3K79me1
    \item H3K79me2
    \item H3K9ac
    \item H3K9me1
    \item H3K9me3
    \item H3T11ph
    \item H4K12ac
    \item H4K20me1
    \item H4K5ac
    \item H4K8ac
    \item H4K91ac
    \item POLR2A
    \item POLR2AphosphoS2
    \item POLR2AphosphoS5
}
}
\keyword{compendium}
\keyword{metagene}
\keyword{profiles}
