% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{hg19_refseq_genes_filtered_granges}
\alias{hg19_refseq_genes_filtered_granges}
\title{Filtered RefSeq gene annotation as GRanges}
\format{A GRanges object.}
\usage{
data(hg19_refseq_genes_filtered_granges)
}
\description{
RefSeq gene annotation data for the CHIC package in GRanges format, filtered
by the gene length and by overlaps.
}
\keyword{annotation}
