% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{refseq_genes_annotated}
\alias{refseq_genes_annotated}
\title{RefSeq gene annotation.}
\format{A data frame with variables:}
\source{

}
\usage{
refseq_genes_annotated
}
\description{
RefSeq gene annotation data for the CHIC package.
}
\examples{


}
\keyword{datasets}
