% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{refseq_genes_tssregion_granges}
\alias{refseq_genes_tssregion_granges}
\title{RefSeq gene transcription start site regions as GRanges}
\format{A GRanges object.}
\usage{
data(refseq_genes_tssregion_granges)
}
\description{
RefSeq gene transcription start site regions for the CHIC package in GRanges
format. Transcription start site regions are defined as ...
}
\keyword{annotation}
