% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chic.data.R
\docType{data}
\name{rf_models}
\alias{rf_models}
\title{Random forest models for ChIP-seq experiment classification}
\format{A list of 7 random forest models.

\itemize{
   \item broadEncode
   \item H3K9Encode
   \item H3K27Encode
   \item H3K36Encode
   \item RNAPol2Encode
   \item sharpEncode
   \item TFmodel
}}
\usage{
data(rf_models)
}
\description{
Random forest models based on ENCODE and Roadmap data for ChIP-seq experiment
classification using the CHIC package.
}
\keyword{classification}
